requirements.txt
Flask==3.0.2
Flask-Session==0.6.0
python-dotenv==1.0.1
requests==2.31.0
msal==1.26.0
